chore: 폰트 오류 수정
