+++ client/src/index.css
@@ -10,28 +10,28 @@
   font-family: 'Pretendard';
   font-weight: 400;
   font-display: swap;
-  src: url('/fonts/pretendard/Pretendard-Regular.woff2') format('woff2');
+  src: url('/font/pretendard/PretendardStd-Regular.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Pretendard';
   font-weight: 500;
   font-display: swap;
-  src: url('/fonts/pretendard/Pretendard-Medium.woff2') format('woff2');
+  src: url('/font/pretendard/PretendardStd-Medium.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Pretendard';
   font-weight: 600;
   font-display: swap;
-  src: url('/fonts/pretendard/Pretendard-SemiBold.woff2') format('woff2');
+  src: url('/font/pretendard/PretendardStd-SemiBold.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Pretendard';
   font-weight: 700;
   font-display: swap;
-  src: url('/fonts/pretendard/Pretendard-Bold.woff2') format('woff2');
+  src: url('/font/pretendard/PretendardStd-Bold.woff2') format('woff2');
 }
 
 /* CSS Variables */
